all working except POST - response & professionalexperience

+++ app/models/talentskills.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, DateTime, ForeignKey, Float
+from sqlalchemy import Column, Integer, DateTime, ForeignKey, Float, Numeric
 from sqlalchemy.orm import relationship
 from .base import Base
 
@@ -8,7 +8,7 @@ class TalentSkill(Base):
     talent_id = Column(Integer, ForeignKey("talents.talent_id"), primary_key=True)
     skill_id = Column(Integer, ForeignKey("skills.skill_id"), primary_key=True)
     proficiency_level = Column(Integer)
-    years_of_experience = Column(Float)
+    years_of_experience = Column(Numeric(4, 2))
     last_used_date = Column(DateTime)
 
     # Relationships
